requirements.txt
livekit-agents>=1.0.0
livekit-plugins-deepgram>=1.0.0
livekit-plugins-anthropic>=1.0.0
livekit-plugins-silero>=1.0.0
livekit>=1.0.0
livekit-api>=1.0.0
httpx>=0.27.0
python-dotenv>=1.0.0
resend>=2.0.0
twilio>=9.0.0
pydantic>=2.0.0
